+++ src/main/java/cn/simplifydb/database/run/read/IsExists.java
@@ -2,6 +2,7 @@ package cn.simplifydb.database.run.read;
 
 import cn.simplifydb.database.base.BaseRead;
 import cn.simplifydb.database.config.DatabaseContextHolder;
+import cn.simplifydb.database.util.JdbcUtil;
 import cn.simplifydb.system.DbLog;
 import cn.simplifydb.util.KeyLock;
 import com.alibaba.druid.sql.ast.SQLLimit;
@@ -90,7 +91,12 @@ public class IsExists<T> extends BaseRead<T> {
             String sql = builder();
             DbLog.getInstance().info(getTransferLog() + getRunSql());
             DataSource dataSource = DatabaseContextHolder.getReadDataSource(tag);
-            return JdbcUtils.executeQuery(dataSource, sql, getParameters());
+            List<Map<String, Object>> list = JdbcUtils.executeQuery(dataSource, sql, getParameters());
+            // 判断是否开启还原
+            if (isUnescapeHtml()) {
+                JdbcUtil.htmlUnescape(list);
+            }
+            return list;
         } finally {
             LOCK.unlock(runClass);
         }