+++ src/main/java/cn/simplifydb/database/run/read/SelectFunction.java
@@ -3,7 +3,9 @@ package cn.simplifydb.database.run.read;
 import cn.jiangzeyin.StringUtil;
 import cn.simplifydb.database.base.BaseRead;
 import cn.simplifydb.database.config.DatabaseContextHolder;
+import cn.simplifydb.database.util.JdbcUtil;
 import cn.simplifydb.database.util.SqlUtil;
+import cn.simplifydb.database.util.Util;
 import cn.simplifydb.system.DbLog;
 import com.alibaba.druid.util.JdbcUtils;
 
@@ -49,13 +51,14 @@ public class SelectFunction<T> extends BaseRead<T> {
             setRunSql(sql);
             DbLog.getInstance().info(getTransferLog() + sql);
             List<Map<String, Object>> list = JdbcUtils.executeQuery(dataSource, sql, getParameters());
-            if (list.size() < 1) {
+            if (Util.checkListMapNull(list)) {
                 return null;
+            }
+            // 判断是否开启还原
+            if (isUnescapeHtml()) {
+                JdbcUtil.htmlUnescape(list);
             }
             Map<String, Object> map = list.get(0);
-            if (map == null) {
-                return null;
-            }
             Collection<Object> collection = map.values();
             return (T) collection.toArray()[0];
         } catch (Exception e) {